+++ tests/test_fmp_get.py
@@ -2,7 +2,11 @@ import pytest
 import httpx
 from pytest_httpx import HTTPXMock
 import polars as pl
-from fmpapi.fmp_get import fmp_get, convert_column_names, convert_column_types, perform_request
+import pandas as pd
+from fmpapi.fmp_get import (
+    fmp_get, convert_column_names, convert_column_types, perform_request, 
+    is_module_available
+)
 
 # Validation tests --------------------------------------------------------
 
@@ -110,3 +114,32 @@ def test_convert_column_types():
     assert df_converted.schema["calendarYear"] == pl.Int32
     assert df_converted.schema["date"] == pl.Date
     assert df_converted.schema["value"] == pl.Int64
+
+# Pandas conversion test ---------------------------------------------------
+
+def test_fmp_get_returns_pandas_data_frame(httpx_mock: HTTPXMock):
+    example_body = {
+        "date": "2024-09-28",
+        "symbol": "ABC",
+        "reportedCurrency": "USD",
+        "cik": "0001234567",
+        "fillingDate": "2024-11-01",
+        "acceptedDate": "2024-11-01 06:01:36",
+        "calendarYear": "2024",
+        "period": "FY",
+        "cashAndCashEquivalents": 67890,
+    }
+
+    httpx_mock.add_response(json=example_body)
+
+    with httpx.Client() as client:
+        result = fmp_get(resource="balance-sheet-statement", symbol="AAPL", to_pandas=True)
+        assert isinstance(result, pd.DataFrame)
+
+def test_is_module_available_true():
+    result = is_module_available("polars")
+    assert result==True
+
+def test_is_module_available_false():
+    result = is_module_available("xxx")
+    assert result==False