+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.2 (2025-03-06)
+
+- Add affiliate link to README
+- Add `to_pandas` parameter to `fmp_get()`
+- Add optional dependencies to allow installation via `pip install fmpapi[pandas]`
+
 ## v0.1.1 (2025-02-02)
 
 - Move dev package dependencies to optional dependency groups